+++ tests/testthat/test-clean_vars.R
@@ -1,4 +1,7 @@
-if (suppressWarnings(
+.runThisTest <- Sys.getenv("RunAllggeffectsTests") == "yes"
+
+if (.runThisTest &&
+    suppressWarnings(
   require("testthat") &&
   require("emmeans") &&
   require("ggeffects")