+++ NEWS.md
@@ -33,6 +33,9 @@
 
 * Fixed issue with `type = "simulate"` for `glmmTMB` models.
 
+* Fixed issue with `gamlss` models that had `random()` function in the
+  model formula.
+
 # ggeffects 1.0.2
 
 ## Breaking changes