+++ NEWS.md
@@ -1,4 +1,4 @@
-# ggeffects 1.0.3
+# ggeffects 1.1.0
 
 ## New supported models